requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
jinja2==3.1.2
websockets==10.4
pyjwt==1.7.1
pydantic==1.10.7
python-multipart==0.0.5